Eight levels climb their own rise, and both open criteria close
A level's way out has ``rise + 1 - deck`` to gain, so one cruising two blocks
over its terrace with a rise of eight spends seven landings on a staircase
against the wall -- which is the owner's "and then stairs" in one number, and
which turned out to be where both remaining defects lived. ECHO SHAFT, THE
BALCONIES, THE SILENCE, WART FIELDS, BASALT FLUES, ROPE BRIDGE, THE WEIR and
THE GROVE now climb to ``rise + 1`` themselves, capped at ``LIFT_MAX``. The
climb goes into each level's first beat, the one that always plays, with
everything after it bumped by the same amount so the shape is preserved -- the
same ast pass the roster was re-authored with.

Re-entry: **one missed jump in 5,762**, on one level visit in fifty-eight,
against 87.9% before the rehash. Unchecked emergency placements **0.41%**
against the 0.4% the criteria ask for and 0.64% an hour ago. Floor contact 2.5
landings a level, worst 5. Fidelity 94.8% placed as authored over 57% designed
content. No-jump walker 48% and 0 of 24 levels end to end.

Frame cost was never a regression: five rounds taken while the test suite ran
in the background read 3.63 ms and a tower/parkour ratio of 1.78; nine rounds
on a quiet machine read **3.00 ms and 1.389**, better than the 1.46 on record.
A run also writes 5,187 cells against 5,780 before the rehash, because the
pedestals went. CLAUDE.md already carried that lesson and it cost an hour
anyway.

docs/RULES.md gains THE DECK RULE at the head of section 3 and the checklist
now runs level_climb and reentry_probe; docs/TOWER.md gains P7 as the first
criterion a level is judged on. 641 passed, 34 skipped.

+++ src/brainrot/scenes/levels/_base.py
@@ -106,7 +106,7 @@ class Level:
                     form = spec.get("form", "full")
                     lift = 0 if form == "floor" else spec.get("lift", 1)
                     surf = lift + FORMS.get(form, 1.0)
-        self.deck = max(1, int(round(surf - 1.0)))
+        self.deck = max(1, round(surf - 1.0))
 
 
 # ---------------------------------------------------------------------------
